Handle white space delimited caseIds in legacy sessions
